Refactor: logo 이미지 파일 변경 및 사용않는 asset 삭제
logo 이미지가 svg와 png 두개 따로 사용하고 있었는데, png 하나로 통일
logo 사용되는 페이지에서 사이즈 조절
사용않는 asset 이미지들 전부 삭제
